Rewriting the Greeting Webpage using components
