+++ src/components/C3div.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import "../styles.css";
+import C3header from './C3header';
+
+function C3div(){
+    return (
+        <div className='challenge3'>
+            <C3header />
+        </div>
+    );
+}
+
+export default C3div;
