+++ src/components/C3header.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import "../styles.css";
+import C3span from './C3span';
+
+function C3header(){
+    return <h1 className='challenge3h1'>Hello Sir, <C3span /></h1>;
+}
+
+export default C3header;
