--- src/components/Heading.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-function Heading(){
-    return <h1> Netflix top 5 shows</h1>;
-}
-
-export default Heading;
