--- src/components/List.jsx
@@ -1,15 +1,0 @@
-import React from 'react';
-
-function List(){
-    return(
-        <ol>
-            <li> Modern Family </li>
-            <li> The Office </li>
-            <li> Friends </li>
-            <li> Family Guy </li>
-            <li> How I met your mother </li>
-        </ol>
-    );
-}
-
-export default List;
